Split up requests and urllib3 unit tests.
